feat: add_order, cancel_order, print_order

+++ src/order_book.hpp
@@ -11,7 +11,7 @@ public:
     void print() const;
 
 private:
-    std::map<double, uint32_t, std::greater<double>> bids_; // highest price first
-    std::map<double, uint32_t>                       asks_; // lowest price first
-    std::unordered_map<uint64_t, Order>              orders_;
+    std::map<int64_t, uint32_t, std::greater<int64_t>> bids_; // highest price first
+    std::map<int64_t, uint32_t>                       asks_; // lowest price first
+    std::unordered_map<uint64_t, Order>              orders_; // order_id : Order
 };
