+++ src/types.hpp
@@ -6,6 +6,6 @@ enum class Side { Bid, Ask };
 struct Order {
     uint64_t id;
     Side     side;
-    double   price;
+    int64_t   price; // scaled integer to avoid floating point error accumulations affecting comparisons
     uint32_t quantity;
 };
